+++ Service2/src/main/java/app/controller/PlaneController.java
@@ -1,5 +1,7 @@
 package app.controller;
 
+import static app.security.SecurityConstants.HEADER_STRING;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,12 +9,14 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import app.entities.Plane;
 import app.forms.PlaneForm;
 import app.repository.PlaneRepository;
+import app.utils.UtilsMethods;
 
 @RestController
 @RequestMapping("/plane")
@@ -26,14 +30,21 @@ public class PlaneController {
 	}
 
 	@PostMapping("/addNew")
-	public ResponseEntity<String> addNewPlane(@RequestBody PlaneForm planeForm) {
+	public ResponseEntity<String> addNewPlane(@RequestBody PlaneForm planeForm,
+			@RequestHeader(value = HEADER_STRING) String token) {
 		try {
 
-			Plane plane = new Plane(planeForm.getName(), planeForm.getCapacity());
+			ResponseEntity<Integer> response = UtilsMethods.sendGet("http://localhost:8080/user/isAdmin/", token);
+			if (response.getBody() == 1) {
 
-			planeRepo.saveAndFlush(plane);
+				Plane plane = new Plane(planeForm.getName(), planeForm.getCapacity());
 
-			return new ResponseEntity<String>("success", HttpStatus.ACCEPTED);
+				planeRepo.saveAndFlush(plane);
+
+				return new ResponseEntity<String>("success", HttpStatus.ACCEPTED);
+			} else {
+				return new ResponseEntity<String>(HttpStatus.FORBIDDEN);
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			return new ResponseEntity<String>(HttpStatus.BAD_REQUEST);
@@ -41,11 +52,18 @@ public class PlaneController {
 	}
 
 	@DeleteMapping("/delete/{x}")
-	public ResponseEntity<Long> deletePlane(@PathVariable Long x) {
+	public ResponseEntity<Long> deletePlane(@PathVariable Long x, @RequestHeader(value = HEADER_STRING) String token) {
 
 		try {
-			planeRepo.deleteById(x);
-			return new ResponseEntity<Long>(x, HttpStatus.ACCEPTED);
+			ResponseEntity<Integer> response = UtilsMethods.sendGet("http://localhost:8080/user/isAdmin/", token);
+			if (response.getBody() == 1) {
+
+				planeRepo.deleteById(x);
+				return new ResponseEntity<Long>(x, HttpStatus.ACCEPTED);
+
+			} else {
+				return new ResponseEntity<Long>(HttpStatus.FORBIDDEN);
+			}
 		} catch (Exception e) {
 			e.printStackTrace();
 			return new ResponseEntity<Long>(HttpStatus.BAD_REQUEST);
